- Desaparecen contructores de MachineType... se usa el builder para todo
git-svn-id: http://svn2.xp-dev.com/svn/tp-tecnicas/trunk/tp-tecnicas@247 c5b30475-30d6-408d-831a-0400e7537d55

+++ src/model/warehouse/Warehouse.java
@@ -142,7 +142,7 @@ public abstract class Warehouse implements MonthlyUpdatable, DailyUpdatable {
 	}
 
 	private double getDefectivePercentage() {
-		return (this.totalProductsMade != 0 ? this.totalDefectiveProductsMade
+		return (this.totalProductsMade != 0 ? (double) this.totalDefectiveProductsMade
 				/ this.totalProductsMade : 0);
 	}
